Updated dashboard.html
I updated the dashboard and essentially made it look like the coins.html page; the user should have their followed coins displayed in a table.  There are two star icons used, the solid star (which would represent that the user is following that coin) and the outline of a star (which would be if the user decides to stop following that coin).  I need help making the stars switch between solid/outlined when pressed.